+++ tests/Integration/parserhooks/GeocodeTest.php
@@ -3,7 +3,7 @@
 namespace Maps\Test;
 
 use DataValues\Geo\Values\LatLongValue;
-use Maps\Elements\Location;
+use Maps\Geocoders\InMemoryGeocoder;
 
 /**
  * @covers MapsGeocode
@@ -17,7 +17,10 @@ class GeocodeTest extends ParserHookTest {
 	 * @see ParserHookTest::getInstance
 	 */
 	protected function getInstance() {
-		return new \MapsGeocode();
+		return new \MapsGeocode( new InMemoryGeocoder( [
+			'New York' => new LatLongValue( 4, 2 ),
+			'Brussels' => new LatLongValue( 2, 3 ),
+		] ) );
 	}
 
 	/**
@@ -26,7 +29,9 @@ class GeocodeTest extends ParserHookTest {
 	public function parametersProvider() {
 		$paramLists = [];
 
-		$paramLists[] = [ 'location' => 'new york city' ];
+		$paramLists[] = [ 'location' => 'New York', '4, 2' ];
+		$paramLists[] = [ 'location' => 'Brussels', '2, 3' ];
+		$paramLists[] = [ 'location' => 'I am a tomato', 'Geocoding failed' ];
 
 		return $this->arrayWrap( $paramLists );
 	}
@@ -40,11 +45,11 @@ class GeocodeTest extends ParserHookTest {
 		$argLists[] = [
 			[
 				'location' => '4,2',
-				'allowcoordinates' => 'yes',
+				'directional' => 'yes',
 			],
 			[
 				'location' => '4,2',
-				'allowcoordinates' => true,
+				'directional' => true,
 			]
 		];