+++ RELEASE-NOTES.md
@@ -7,6 +7,8 @@ Under development
 
 * Added persistent geocoding cache (by hidrarga)
 * Removed `egMapsUserGeoOverrides` setting
+* Removed `geoservice` parameter from the `#geocode` parser function
+* Removed `allowcoordinates` parameter from the `#geocode` parser function
 * Removed `service` parameter from the `geocode` API module
 * Removed support for the GeocoderUS geocoding service as it appears to have shut down